Corrijo angulos & transformo arcotangente

+++ Server/Model/game.h
@@ -49,7 +49,7 @@ class Game {
     void _reduceBullets(const ConnectionId id);
     void _move(const ConnectionId id);
     void _deletePlayer(ConnectionId id);
-    bool is_player_target(int pos_x_attacker, int pos_y_attacker, int vision_angle_attacker, \
+    bool is_player_target(int pos_x_attacker, int pos_y_attacker, float vision_angle_attacker, \
 int pos_x_other_player,int pos_y_other_player);
 
     public:
